Mocked gramar advert client

+++ GPTipsBot/Dtos/UserProfileDto.cs
@@ -2,7 +2,7 @@
 {
     public string FirstName { get; set; }
     public string? LastName { get; set; }
-    public long Stars { get; set; }
+    public double Stars { get; set; }
     public long Images { get; set; }
     public long ImageTexts { get; set; }
     public long GptRequests { get; set; }
